Refactor: split Response. Delete unused files
