+++ src/server/Response.hpp
@@ -79,50 +79,12 @@ std::ostream &operator<<(std::ostream &, Response const &);
 class Request;
 class ServerConfig;
 
+enum PathType { IS_DIR, IS_FILE, PATH_ERROR };
+
 class ServerResponse {
 public:
-  static std::string find_file_type(const std::string &path);
-  static std::string get_mime_type_for_extension(const std::string &ext);
-  static Response http_response(
-      const Request *request, const ClientSession *client,
-      const std::map<std::string, std::string> &mime_type,
-      Session *session, char **envp);
-  static Result<Void> register_cgi(
-      const Request &request, const RouteRule_CGI &rule, EPoll *epoll,
-      std::map<std::string, std::string> const &cgi_interpreters,
-      std::map<FileDescriptor const *,
-               std::pair<FileDescriptor const *, CgiDelegate *> > &cgis,
-      FileDescriptor const *client_fd, char **envp);
-
-private:
-  enum PathType { IS_DIR, IS_FILE, PATH_ERROR };
-  static Result<int> check_path_type(const std::string &path);
-  static Target resolve_target(const RouteRule *rule,
-                               const ServerConfig *config,
-                               const Request *request, char **envp);
-  static Result<std::string> compute_etag(const std::string &path);
-  static Result<std::string> get_last_modified(const std::string &path);
-  static Response error_response(const ServerConfig *config,
-                                 const RouteRule *rule,
-                                 Response::StatusCode error_code, char **envp);
   static Result<std::string> make_autoindex_page(const std::string &real_path,
                                                  const std::string &req_uri, DIR *dir);
-  static std::string extract_boundary(const std::string &content_type);
-  static std::size_t parse_multipart_part(
-                       const std::string &body, const std::string &boundary,
-                       std::size_t start_pos, std::string &out_filename,
-                       std::string &out_fieldname, std::string &out_data);
-
-  static Response delete_method(const Target &target, Response response,
-                                const ServerConfig *config,
-                                const RouteRule *rule, char **envp);
-  static Response post_method(const Target &target, Response response,
-                              const ClientSession *client,
-                              const RouteRule *rule, const Request *request,
-                              Session *session, char **envp);
-  static Response get_method(Target target, Response response,
-                             const ServerConfig *config, const RouteRule *rule,
-                             const Request *request, char **envp);
 };
 
 class DefaultError {
