+++ src/server/ResponseErrors.hpp
@@ -1,0 +1,19 @@
+#ifndef RESPONSE_ERRORS_HPP
+#define RESPONSE_ERRORS_HPP
+
+#include "Response.hpp"
+#include "../config/ServerConfig.hpp"
+
+struct Target;
+
+class ResponseErrors {
+public:
+  static Response error_response(const ServerConfig *config,
+                                 const RouteRule *rule,
+                                 Response::StatusCode error_code, char **envp);
+  static Target resolve_target(const RouteRule *rule,
+                               const ServerConfig *config,
+                               const Request *request, char **envp);
+};
+
+#endif
