--- www-files/*/d1.txt
@@ -1,1 +1,0 @@
-hello world from D1
