--- www-files/*/d2.txt
@@ -1,1 +1,0 @@
-hello world split into many smaller pieces today
