--- www-files/*/d3.bin
@@ -1,1 +1,0 @@
-0000000000000001000000000000000200000000000000030000000000000004000000000000000500000000000000060000000000000007000000000000000800000000000000090000000000000010000000000000001100000000000000120000000000000013000000000000001400000000000000150000000000000016000000000000001700000000000000180000000000000019000000000000002000000000000000210000000000000022000000000000002300000000000000240000000000000025000000000000002600000000000000270000000000000028000000000000002900000000000000300000000000000031000000000000003200000000000000330000000000000034000000000000003500000000000000360000000000000037000000000000003800000000000000390000000000000040000000000000004100000000000000420000000000000043000000000000004400000000000000450000000000000046000000000000004700000000000000480000000000000049000000000000005000000000000000510000000000000052000000000000005300000000000000540000000000000055000000000000005600000000000000570000000000000058000000000000005900000000000000600000000000000061000000000000006200000000000000630000000000000064
